rename everythign to iris_rails, copying helper and iris_rails.rb files now
