+++ bin/console
@@ -1,5 +1,4 @@
 #!/usr/bin/env ruby
-
 require "bundler/setup"
 require "iris"
