+++ lib/generators/iris/install_generator.rb
@@ -17,7 +17,10 @@ module Iris
         directory "app/views/iris_rails", "app/views/iris_rails"
 
         # Create helper
-        directory "app/helpers/meta_tags_helper", "app/helpers/meta_tags_helper"
+        copy_file "app/helpers/meta_tags_helper", "app/helpers/meta_tags_helper"
+
+        # Create helper
+        copy_file "../lib/iris_rails.rb", "../lib/iris_rails.rb"
 
         # Create default layout
         # directory "app/views/layouts/iris", "app/views/layouts/iris"
