+++ lib/generators/lib/iris_rails.rb
