+++ lib/generators/templates/config/initializers/iris_rails.rb
@@ -1,2 +1,2 @@
-require 'iris-rails'
+require 'iris_rails'
 IrisRails::api_token = ENV['IRIS_TOKEN']
